fix: 報表資訊錯誤時卡 loading
